LIBFCREPO-896. Initial implementation of IpMapperFilter
Largely followed the existing code in
"edu.umd.lib.tomcat.valves.IPAddressMapper", including the following:

* relying on a "mappingFile" (specified using a "mappingFile" parameter
  in the "web.xml" file).
* the HTTP header containing the allowed mappings defaults to
  "X-Auth-IP-Mapper" and is configurable via a "headerName" parameter
  in the "web.xml" file.
* The code currently only supports IPv4 addresses

Decided to add a dependency to the
"org.springframework.security:spring-security-web" library to allow use
of the "IpAddressMatcher" class. This is in place of using the
"SubnetUtils/SubnetInfo" from Apache Commons, as in "ip-mapper-valve",
because the "IpAddressMatcher" class can match single IP addresses
(with or without the "/32" CIDR" suffix), as well as subnets without
any special handling.

https://issues.umd.edu/browse/LIBFCREPO-896