LIBFCREPO-903. Pin "tomcat" base image to image using Java 8v265
The latest version of the "tomcat:7.0.106-jdk8-openjdk-buster" Docker
image uses OpenJDK v8u272, which has a bug that causes an exception
to be thrown when attempting to contact the LDAP server. This is likely
a regression from https://bugs.openjdk.java.net/browse/JDK-8214440

Pinning the "tomcat" base image using a sh256 digest, so that an
earlier version of the image using OpenJDK v8u256 is used.

https://issues.umd.edu/browse/LIBFCREPO-903