LIBFCREPO-1001. Implement a "store-and-forward" message broker.
Implemented as an embedded ActiveMQ instance, listening on localhost:61616 for OpenWire connections. Configured as a static bridge of the "fedora" queue to the external message broker at the location given in ACTIVEMQ_URL. Added a /var/activemq volume to the Docker image for use as the persistent store for the queue.

https://issues.umd.edu/browse/LIBFCREPO-1001.