LIBFCREPO-1051. Bumped Tomcat base Docker image to 7.0.109
Bumped Tomcat base Docker image to v7.0.109. This version has a fix
for the LDAP issue that was causing us to "pin" the Tomcat Docker image
to a specific "7.0.106-jdk8-openjdk-buster" version.

The Docker image is still "pinned" using an SHA256 hash, to ensure that
the same base image is always used.

Also switch to the "slim" Docker image to save disk space.

https://issues.umd.edu/browse/LIBFCREPO-1051