LIBFCREPO-1051. Use Tomcat base image for "linux/amd64" architecture
Modified the Tomcat base image to retrieve the image for the
"linux/amd64" architecture, instead of the "linux/arm64/v8"
architecture.

This was needed because the image was causing the following error when
run in Kubernetes

```
standard_init_linux.go:228: exec user process caused: exec format error
```

The Docker image is still "pinned" using an SHA256 hash, to ensure that
the same base image is always used.

https://issues.umd.edu/browse/LIBFCREPO-1051