LIBFCREPO-1518. Removed "basic-auth.properties" file and references
The "basic-auth.properties" file was used to support the Loris IIIF
image server that is no longer being used, and was removed from
the "k8s-fcrepo" configuration.

Removing the Spring configuration that relies on the
"basic-auth.properties" file and the "CREDENTIALS_FILE" so that
the application will start in Kubernetes.

Left the "BasicAuthFilter" Java class (with an explanatory note) in
case it is needed in the future.

https://umd-dit.atlassian.net/browse/LIBFCREPO-1518